feat : added new auxiliary function to setup authentication tokens in test files
